Added source tag to fitbit output update props.conf & inputs.conf
In order to ensure that different sources are identified and
processed correctly I added a field to the FitbitOutput struct
that lists the source and created a method to make sure that it is
always set.

Additionally I updated `default/inputs.conf` with default input
settings for a google and a fitbit input. Then updated
`default/props.conf` to match and do JSON and date extraction.